Work 硬件仿真测试环境合集版37_2011_8_24_1_x_5:(1)将ADC数据输出改为整形输出，格式与施工程序输出文件统一。
git-svn-id: https://cxm-hitech-matrix428.googlecode.com/svn/trunk@185 5d2555b2-77f6-cb3c-1dbd-c1bb7278d601

+++ 硬件仿真测试环境合集版37_CMap改为Map_2011_5_30_1/MatrixTestSoft/MatrixTestSoft/TabADCDataShow.h
@@ -24,7 +24,7 @@ private:
 	void ProcessMessages(void);
 public:
 	// ADC数据界面显示缓冲区指针
-	double* m_pADCDataShow;
+	unsigned int* m_pADCDataShow;
 public:
 	afx_msg void OnTimer(UINT_PTR nIDEvent);
 };
